+++ Tests/testFmmDemonstration.cpp
@@ -135,7 +135,7 @@ int main(int argc, char ** argv){
 
     const int NbLevels = FParameters::getValue(argc,argv,"-h", 5);
     const int SizeSubLevels = FParameters::getValue(argc,argv,"-sh", 3);
-    const long NbPart = FParameters::getValue(argc,argv,"-pn", 20L);
+    const int NbPart = FParameters::getValue(argc,argv,"-pn", 20L);
     FTic counter;
 
     //////////////////////////////////////////////////////////////////////////////////
