+++ Tests/testOctreeRearrange.cpp
@@ -45,7 +45,7 @@ int main(int argc, char ** argv){
 
     const int NbLevels          = FParameters::getValue(argc,argv,"-h", 7);
     const int SizeSubLevels     = FParameters::getValue(argc,argv,"-sh", 3);
-    const long NbPart           = FParameters::getValue(argc,argv,"-nb", 2000000);
+    const int NbPart           = FParameters::getValue(argc,argv,"-nb", 2000000);
     const FReal FRandMax        = FReal(RAND_MAX);
 
     FTic counter;
